# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation de l'API Laravel",
        "description": "Implémentation de Swagger avec L5-Swagger.",
        "contact": {
            "name": "KANHONOU AUSCENCE",
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de l'API"
        }
    ],
    "paths": {
...
        "/api/user/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Connexion d'un utilisateur",
                "operationId": "9f6f4aca790a5229fb638e798623c6e0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Connexion effectué avec succès. "
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Enregistre un nouvel utilisateur",
                "operationId": "d03ee3a44bdd6f0025efd14c5c76b5ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "Depute"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succès. (Retourne généralement le cookie de session Sanctum et un statut 201)."
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}